move evaluate parsers
