+++ elk/__main__.py
@@ -6,11 +6,10 @@ from argparse import ArgumentParser
 from contextlib import nullcontext, redirect_stdout
 
 from elk.evaluate.evaluate import evaluate
-from elk.evaluate.parser import get_evaluate_parser
 from elk.files import args_to_uuid, elk_cache_dir
 from elk.list_runs import list_runs
 
-from .argparsers import get_extraction_parser, get_training_parser
+from .argparsers import get_extraction_parser, get_training_parser, get_evaluate_parser
 from .extraction.extraction_main import run as run_extraction
 from .training.train import train
 
@@ -19,47 +18,6 @@ from .training.train import train
 def sweep(args):
     """
     Train and evaluate ccs and lr model on a set of datasets.
-    """
-
-    # extract and train
-    names = []
-    for model in args.models:
-        args.model = model
-        for dataset in args.datasets:
-            args.dataset = dataset
-            args.name = args_to_uuid(args)
-            # TODO: Set missing args...
-            run_extraction(args)
-            train(args)
-
-            names.append(args.name)
-
-    for name in names:
-        args.reporters = name
-        args.hidden_states = names
-        evaluate(args)
-
-
-# TODO: Move function to a better place...
-def get_sweep_parser():
-    parser = ArgumentParser(add_help=False)
-    add_sweep_args(parser)
-    return parser
-
-
-# TODO: Move function to a better place...
-def add_sweep_args(parser):
-    parser.add_argument("--models", nargs="+", type=str, help="Model to sweep over.")
-    parser.add_argument(
-        "--datasets", nargs="+", type=str, help="Dataset to sweep over."
-    )
-
-
-# TODO: Move function to a better place...
-def sweep(args):
-    """
-    Train and evaluate ccs and lr model on a set of datasets.
-    Example: elk sweep --models google/t5-v1_1-base microsoft/deberta-v2-xxlarge-mnli --datasets imdb amazon_polarity --max-examples 100
     """
 
     # extract and train
